created cluster-test.ts
