+++ projects/safe/src/lib/components/widgets/map/map.component.ts
@@ -9,7 +9,7 @@ import { SafeUnsubscribeComponent } from '../../utils/unsubscribe/unsubscribe.co
 import 'leaflet.markercluster';
 import 'leaflet.control.layers.tree';
 import { complexGeoJSON, cornerGeoJSON, pointGeoJSON } from './geojson-test';
-import { generateGeoJSONPoints } from './util-test';
+import { generateClusterLayer } from './cluster-test';
 
 // Declares L to be able to use Leaflet from CDN
 declare let L: any;
@@ -247,28 +247,7 @@ export class SafeMapComponent
       },
     };
 
-    // Clusters
-    const clusterGroup = L.markerClusterGroup({
-      zoomToBoundsOnClick: false,
-    });
-    clusterGroup.on('clusterclick', (event: any) => {
-      const children = event.layer.getAllChildMarkers();
-      let popupContent = 'test popup';
-      children.forEach((child: any) => {
-        console.log(child);
-        popupContent += ' and new test';
-      });
-      L.popup()
-        .setLatLng(event.latlng)
-        .setContent(popupContent)
-        .openOn(this.map);
-    });
-    const clusterLayer = L.geoJSON(generateGeoJSONPoints(200), {
-      onEachFeature: (feature: any, layer: any) => {
-        layer.bindPopup('point popup');
-      },
-    });
-    clusterGroup.addLayer(clusterLayer);
+    const clusterGroup = generateClusterLayer(this.map, L);
     this.map.addLayer(clusterGroup);
 
     this.overlays = {
